+++ modules/vpc/versions.tf
@@ -20,18 +20,18 @@ terraform {
   required_providers {
     google = {
       source  = "hashicorp/google"
-      version = ">= 4.64, < 6"
+      version = ">= 4.64, < 7"
     }
     google-beta = {
       source  = "hashicorp/google-beta"
-      version = ">= 4.64, < 6"
+      version = ">= 6.13, < 7"
     }
   }
 
   provider_meta "google" {
-    module_name = "blueprints/terraform/terraform-google-network:vpc/v9.1.0"
+    module_name = "blueprints/terraform/terraform-google-network:vpc/v10.0.0"
   }
   provider_meta "google-beta" {
-    module_name = "blueprints/terraform/terraform-google-network:vpc/v9.1.0"
+    module_name = "blueprints/terraform/terraform-google-network:vpc/v10.0.0"
   }
 }