+++ test/setup/iam.tf
@@ -16,6 +16,7 @@
 
 locals {
   int_required_roles = [
+    "roles/compute.instanceAdmin",
     "roles/compute.networkAdmin",
     "roles/compute.securityAdmin",
     "roles/iam.serviceAccountUser",
@@ -25,6 +26,7 @@ locals {
     "roles/resourcemanager.tagAdmin",
     "roles/iam.serviceAccountAdmin",
     "roles/compute.orgFirewallPolicyAdmin",
+    "roles/networkconnectivity.hubAdmin",
   ]
 }